fixed toasts and state
