+++ assets/index.html
@@ -35,20 +35,6 @@
     <div id="app"></div>
     <div id="toasts"></div>
 
-    <script type="text/x-template" id="message-template">
-      <div class="toast" data-delay="10000" v-bind:data-autohide="true">
-        <div class="toast-header">
-          <strong class="mr-auto" v-if="item.type != 'warn'"><font-awesome-icon class="text-danger" icon="exclamation-triangle" /> Error</strong>
-          <strong class="mr-auto" v-if="item.type == 'warn'"><font-awesome-icon class="text-warning" icon="exclamation-triangle" /> Warning</strong>
-          <small v-if="item.status">HTTP {{item.status}}</small>
-          <button type="button" class="ml-2 mb-1 close" data-dismiss="toast" aria-label="Close">
-            <span aria-hidden="true">&times;</span>
-          </button>
-        </div>
-        <div class="toast-body">{{item.message}}</div>
-      </div>
-    </script>
-
     <footer class="pt-4 pb-2 pt-md-5 pb-md-4 mt-3 mt-md-5 border-top">
       <div class="container">
         <div class="row">
