+++ assets/js/components/MessageToast.vue
@@ -1,13 +1,19 @@
 <template>
-	<div aria-atomic="true" style="position: absolute; top: 4rem; right: 0.5rem">
-		<!-- Position it -->
-		<MessageToast
-			v-for="item in items"
-			v-bind:item="item"
-			:id="'message-id-' + item.id"
-			:key="item.id"
-		>
-		</MessageToast>
+	<div class="toast" data-delay="10000" v-bind:data-autohide="true">
+		<div class="toast-header">
+			<strong class="mr-auto" v-if="item.type != 'warn'"
+				><font-awesome-icon class="text-danger" icon="exclamation-triangle" /> Error</strong
+			>
+			<strong class="mr-auto" v-if="item.type == 'warn'"
+				><font-awesome-icon class="text-warning" icon="exclamation-triangle" />
+				Warning</strong
+			>
+			<small v-if="item.status">HTTP {{ item.status }}</small>
+			<button type="button" class="ml-2 mb-1 close" data-dismiss="toast" aria-label="Close">
+				<span aria-hidden="true">&times;</span>
+			</button>
+		</div>
+		<div class="toast-body">{{ item.message }}</div>
 	</div>
 </template>
